rizzipe_backend/requirements.txt
Flask==2.1.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Flask-RESTful==0.3.9
requests==2.26.0
Pillow==8.4.0
flask_script==2.0.5
python-dotenv==0.19.2
honcho==1.0.1
asyncio==3.4.3
openai==0.27.8
torch==2.0.1
torchvision==0.15.2
pandas~=1.5.3
supabase==1.0.3
